Added unittest test_more_voices
Changed filename of main file: minus sign not allowed in python file names and therefore only importable using workarounds
Made args global to fix the todo that args is being passed too deep into the call stack
Created main method to be able to call the script from outside after import
Created unittest for SATB, SSAA and SSAATTBB
Added Abendlied-Rheinberger.mxl to implicitely test mxl midi conversion and SSAATTBB

+++ tests/test-integration.py
@@ -11,7 +11,8 @@ def audio_length(mp3_filepath):
 
 class TestIntegration(unittest.TestCase):
     def setUp(self):
-        shutil.rmtree(output_directory)
+        if(os.path.exists(output_directory)):
+            shutil.rmtree(output_directory)
 
     def test_conversion_of_midi_with_separate_tempo_map(self):
         midi_file_path = "./tests/fixtures/Schumann-op67-4.mid"
